Azure/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
cloudpickle
# numpy
# scikit-learn
# scipy
# implicit
# psutil
# mlflow
# pandas

